+++ src/math/vector/vec3/vec3_mult.c
@@ -6,7 +6,7 @@
 /*   By: cybattis <cybattis@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/13 19:58:54 by cybattis          #+#    #+#             */
-/*   Updated: 2022/01/14 15:46:49 by cybattis         ###   ########.fr       */
+/*   Updated: 2022/02/03 19:28:42 by cybattis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ t_vec3	*vec3_multi(t_vec3 *self, long n)
 	return (self);
 }
 
-t_vec3	*vec3_multf(t_vec3 *self, double n)
+t_vec3	*vec3_multf(t_vec3 *self, float n)
 {
 	self->x *= n;
 	self->y *= n;
